requirements.txt
Django==2.2.27
djangorestframework==3.11.0
pytz==2021.3
sqlparse==0.4.2